refactor(rechnung): Betraege durchgehend als BigDecimal
Die Rechnungsgenerierung rechnete mit double, die Debitorenkontrolle und die
Nebenkostenabrechnung dagegen mit BigDecimal. Der Uebergang zur Forderung
musste deshalb konvertieren, und double kann Rappenbetraege nicht exakt
darstellen. Umgestellt wird die Rechnungsgenerierung - damit ist BigDecimal
die Regel und nicht mehr die Nebenkostenabrechnung die Ausnahme.

- RechnungDTO und TarifZeileDTO fuehren Betraege, Mengen und Preise als
  BigDecimal. tarif.preis (NUMERIC(10,5)) wird unveraendert uebernommen statt
  ueber doubleValue(); der Zeilenbetrag ist das exakte Produkt.
- roundTo5Rappen nimmt und liefert BigDecimal und rundet mit HALF_UP, also
  von Null weg und damit symmetrisch fuer negative Betraege. Das frueher
  verwendete Math.round(x * 20.0) / 20.0 rundete -0.5 gegen +unendlich.
- Der Endbetrag traegt zwei Nachkommastellen und geht ohne Umrechnung in
  DebitorService.upsertFromRechnung; das BigDecimal.valueOf(...).setScale(2)
  im RechnungController entfaellt.
- Die Rundungsregel bleibt unveraendert: gerundet wird nur der Endbetrag, auf
  5 Rappen. Die Betraege auf der Rechnung aendern sich nicht.
- Unveraendert double bleiben die kWh-Summen aus dem MesswerteRepository -
  sie kommen so aus der Datenbank und werden direkt auf ganze kWh gerundet.

PDF: rechnung.jrxml deklariert menge, preis und betrag als
java.math.BigDecimal. Die bedingte Rundungszeile prueft nun
getRundung().signum() != 0 statt Math.abs(...) > 0.001 - Math.abs hat keine
BigDecimal-Ueberladung, und die Schwelle war ohnehin nur ein Behelf gegen die
Unschaerfe von double.

Tests: JasperTemplateCompileTest fuellt das Rechnungs-Template neu auch mit
Daten und exportiert ein PDF. Ein Template kompiliert auch mit Feldtypen, die
nicht zur Bean passen; der ClassCastException kommt erst beim Fuellen und trat
bei dieser Umstellung genau so auf. Der Test kompiliert aus dem .jrxml und
nicht ueber RechnungPdfService, weil dessen rechnung.jasper erst in der Phase
prepare-package entsteht und bei mvn test vom vorherigen Lauf stammt.

In RechnungServiceTest vergleicht ein Helfer assertBetrag ueber compareTo.
Alle 53 Erwartungswerte stimmen exakt - die bisherigen Toleranzen von 0.01
fingen allein die Unschaerfe von double auf. Vier zweiargumentige Assertions
haetten nach der Umstellung stillschweigend fehlgeschlagen, weil
assertEquals(Object, Object) einen Double nie einem BigDecimal gleich findet.

Keine Migration, kein Frontend: die Entities waren schon NUMERIC, und
BigDecimal serialisiert wie double als JSON-Zahl.

+++ backend-service/src/main/java/ch/nacht/dto/TarifZeileDTO.java
@@ -2,27 +2,33 @@ package ch.nacht.dto;
 
 import ch.nacht.entity.TarifTyp;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 /**
  * DTO for a single tariff line on an invoice.
  * Multiple lines can exist per tariff type if tariffs change within the invoice period.
+ *
+ * <p><b>Betraege sind {@link BigDecimal}</b> - wie in {@code Debitor} und in der
+ * Nebenkostenabrechnung. {@code preis} wird unveraendert aus {@code Tarif.preis}
+ * ({@code NUMERIC(10,5)}) uebernommen, {@code betrag} ist das exakte Produkt
+ * {@code menge x preis}; gerundet wird erst der Endbetrag der Rechnung auf 5 Rappen.
  */
 public class TarifZeileDTO {
 
     private String bezeichnung;
     private LocalDate von;
     private LocalDate bis;
-    private double menge;      // kWh (rounded)
-    private double preis;      // CHF/kWh
-    private double betrag;     // CHF
+    private BigDecimal menge;      // kWh (rounded)
+    private BigDecimal preis;      // CHF/kWh
+    private BigDecimal betrag;     // CHF
     private TarifTyp typ;
     private String mengeneinheit; // "kWh" for ZEV/VNB, "Monate" for GRUNDGEBUEHR
 
     public TarifZeileDTO() {
     }
 
-    public TarifZeileDTO(String bezeichnung, LocalDate von, LocalDate bis, double menge, double preis, double betrag, TarifTyp typ) {
+    public TarifZeileDTO(String bezeichnung, LocalDate von, LocalDate bis, BigDecimal menge, BigDecimal preis, BigDecimal betrag, TarifTyp typ) {
         this.bezeichnung = bezeichnung;
         this.von = von;
         this.bis = bis;
@@ -33,7 +39,7 @@ public class TarifZeileDTO {
         this.mengeneinheit = "kWh";
     }
 
-    public TarifZeileDTO(String bezeichnung, LocalDate von, LocalDate bis, double menge, double preis, double betrag, TarifTyp typ, String mengeneinheit) {
+    public TarifZeileDTO(String bezeichnung, LocalDate von, LocalDate bis, BigDecimal menge, BigDecimal preis, BigDecimal betrag, TarifTyp typ, String mengeneinheit) {
         this.bezeichnung = bezeichnung;
         this.von = von;
         this.bis = bis;
@@ -68,27 +74,27 @@ public class TarifZeileDTO {
         this.bis = bis;
     }
 
-    public double getMenge() {
+    public BigDecimal getMenge() {
         return menge;
     }
 
-    public void setMenge(double menge) {
+    public void setMenge(BigDecimal menge) {
         this.menge = menge;
     }
 
-    public double getPreis() {
+    public BigDecimal getPreis() {
         return preis;
     }
 
-    public void setPreis(double preis) {
+    public void setPreis(BigDecimal preis) {
         this.preis = preis;
     }
 
-    public double getBetrag() {
+    public BigDecimal getBetrag() {
         return betrag;
     }
 
-    public void setBetrag(double betrag) {
+    public void setBetrag(BigDecimal betrag) {
         this.betrag = betrag;
     }
